Deepfake_backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
opencv-python>=4.8.0
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
pillow>=10.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
werkzeug>=2.3.0
tqdm>=4.65.0
matplotlib>=3.7.0
datasets>=2.10.0
